requirements.txt
kafka-python
hdfs
pandas
streamlit
requests

# pip install -r requirements.txt
